Extract each plugin to its own source file
