+++ src/main/scala/interplay/PlayNoPublishBase.scala
@@ -1,0 +1,21 @@
+package interplay
+
+import sbt._
+import sbt.Keys._
+import sbt.Resolver
+import com.jsuereth.sbtpgp.PgpKeys
+
+/**
+ * 
+ */
+object PlayNoPublishBase extends AutoPlugin {
+  override def trigger = noTrigger
+  override def requires = PlayBuildBase
+
+  override def projectSettings = Seq(
+    PgpKeys.publishSigned := {},
+    publish := {},
+    publishLocal := {},
+    publishTo := Some(Resolver.file("no-publish", crossTarget.value / "no-publish"))
+  )
+}
