+++ src/main/scala/interplay/PlaySonatypeBase.scala
@@ -1,0 +1,19 @@
+package interplay
+
+import sbt.AutoPlugin
+import sbt.Keys.publishTo
+import xerial.sbt.Sonatype
+import xerial.sbt.Sonatype.autoImport.{ sonatypePublishToBundle, sonatypeProfileName }
+
+/**
+ * Base plugin for all projects that publish to sonatype
+ */
+object PlaySonatypeBase extends AutoPlugin {
+  override def trigger = noTrigger
+  override def requires = Sonatype
+
+  override def projectSettings = Seq(
+    sonatypeProfileName := "com.typesafe",
+    publishTo := sonatypePublishToBundle.value,
+  )
+}
