+++ src/main/scala/interplay/Versions.scala
@@ -1,0 +1,12 @@
+package interplay
+
+object ScalaVersions {
+  val scala210 = "2.10.7"
+  val scala212 = "2.12.10"
+  val scala213 = "2.13.1"
+}
+
+object SbtVersions {
+  val sbt013 = "0.13.18"
+  val sbt10 = "1.2.8"
+}
